pyproject.toml
[project]
name = "secure-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx>=0.2.4",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
]
